+++ include/main.h
@@ -56,6 +56,17 @@ struct settings {
 */
 typedef struct settings Settings;
 
+struct timings {
+    int nmeaTook;
+    int nmeaRest;
+    int webTook;
+    int webRest;
+    int bgTook;
+    int bgRest;
+};
+
+typedef struct timings Timings;
+
 /**
  * @brief Save current settings to the SPIFFS.
  * @return True if succeeded
@@ -85,3 +96,5 @@ void crash();
  * @return True if present
 */
 bool getSDcardStatus();
+
+void vBackgroundTasks(void * pvParameters);